requirements.txt
azure-functions>=1.20.0
azure-identity==1.17.1
azure-storage-blob==12.19.1

